fix: 유니온 컴포넌트 폴더 분리
