--- src/components/user/union/UnionIcon.js
@@ -1,69 +1,0 @@
-import novice_rank1_icon from "../../../assets/union/grade/novice/novice_rank1.png";
-import novice_rank2_icon from "../../../assets/union/grade/novice/novice_rank2.png";
-import novice_rank3_icon from "../../../assets/union/grade/novice/novice_rank3.png";
-import novice_rank4_icon from "../../../assets/union/grade/novice/novice_rank4.png";
-import novice_rank5_icon from "../../../assets/union/grade/novice/novice_rank5.png";
-
-import veteran_rank1_icon from "../../../assets/union/grade/veteran/veteran_rank1.png";
-import veteran_rank2_icon from "../../../assets/union/grade/veteran/veteran_rank2.png";
-import veteran_rank3_icon from "../../../assets/union/grade/veteran/veteran_rank3.png";
-import veteran_rank4_icon from "../../../assets/union/grade/veteran/veteran_rank4.png";
-import veteran_rank5_icon from "../../../assets/union/grade/veteran/veteran_rank5.png";
-
-import master_rank1_icon from "../../../assets/union/grade/master/master_rank1.png";
-import master_rank2_icon from "../../../assets/union/grade/master/master_rank2.png";
-import master_rank3_icon from "../../../assets/union/grade/master/master_rank3.png";
-import master_rank4_icon from "../../../assets/union/grade/master/master_rank4.png";
-import master_rank5_icon from "../../../assets/union/grade/master/master_rank5.png";
-
-import grand_master_rank1_icon from "../../../assets/union/grade/grand_master/grand_master_rank1.png";
-import grand_master_rank2_icon from "../../../assets/union/grade/grand_master/grand_master_rank2.png";
-import grand_master_rank3_icon from "../../../assets/union/grade/grand_master/grand_master_rank3.png";
-import grand_master_rank4_icon from "../../../assets/union/grade/grand_master/grand_master_rank4.png";
-import grand_master_rank5_icon from "../../../assets/union/grade/grand_master/grand_master_rank5.png";
-
-import supreme_rank1_icon from "../../../assets/union/grade/supreme/supreme_rank1.png";
-import supreme_rank2_icon from "../../../assets/union/grade/supreme/supreme_rank2.png";
-import supreme_rank3_icon from "../../../assets/union/grade/supreme/supreme_rank3.png";
-import supreme_rank4_icon from "../../../assets/union/grade/supreme/supreme_rank4.png";
-import supreme_rank5_icon from "../../../assets/union/grade/supreme/supreme_rank5.png";
-
-const UnionIcons = {
-  novice: [
-    novice_rank1_icon,
-    novice_rank2_icon,
-    novice_rank3_icon,
-    novice_rank4_icon,
-    novice_rank5_icon,
-  ],
-  veteran: [
-    veteran_rank1_icon,
-    veteran_rank2_icon,
-    veteran_rank3_icon,
-    veteran_rank4_icon,
-    veteran_rank5_icon,
-  ],
-  master: [
-    master_rank1_icon,
-    master_rank2_icon,
-    master_rank3_icon,
-    master_rank4_icon,
-    master_rank5_icon,
-  ],
-  grand_master: [
-    grand_master_rank1_icon,
-    grand_master_rank2_icon,
-    grand_master_rank3_icon,
-    grand_master_rank4_icon,
-    grand_master_rank5_icon,
-  ],
-  supreme: [
-    supreme_rank1_icon,
-    supreme_rank2_icon,
-    supreme_rank3_icon,
-    supreme_rank4_icon,
-    supreme_rank5_icon,
-  ],
-};
-
-export default UnionIcons;
