+++ src/components/user/union/unionArtifact/UnionArtifactEffect.jsx
