+++ src/components/user/union/unionArtifact/UnionArtifactIcon.js
@@ -1,0 +1,39 @@
+import artifect_0_grade1_icon from "../../../../assets/union/artifact/artifect_0_grade1_icon.png";
+import artifect_0_grade5_icon from "../../../../assets/union/artifact/artifect_0_grade5_icon.png";
+
+import artifect_1_grade1_icon from "../../../../assets/union/artifact/artifect_1_grade1_icon.png";
+import artifect_1_grade5_icon from "../../../../assets/union/artifact/artifect_1_grade5_icon.png";
+
+import artifect_2_grade1_icon from "../../../../assets/union/artifact/artifect_2_grade1_icon.png";
+import artifect_2_grade5_icon from "../../../../assets/union/artifact/artifect_2_grade5_icon.png";
+
+import artifect_3_grade1_icon from "../../../../assets/union/artifact/artifect_3_grade1_icon.png";
+import artifect_3_grade5_icon from "../../../../assets/union/artifact/artifect_3_grade5_icon.png";
+
+import artifect_4_grade1_icon from "../../../../assets/union/artifact/artifect_4_grade1_icon.png";
+import artifect_4_grade5_icon from "../../../../assets/union/artifact/artifect_4_grade5_icon.png";
+
+import artifect_5_grade1_icon from "../../../../assets/union/artifact/artifect_5_grade1_icon.png";
+import artifect_5_grade5_icon from "../../../../assets/union/artifact/artifect_5_grade5_icon.png";
+
+import artifect_6_grade1_icon from "../../../../assets/union/artifact/artifect_6_grade1_icon.png";
+import artifect_6_grade5_icon from "../../../../assets/union/artifact/artifect_6_grade5_icon.png";
+
+import artifect_7_grade1_icon from "../../../../assets/union/artifact/artifect_7_grade1_icon.png";
+import artifect_7_grade5_icon from "../../../../assets/union/artifact/artifect_7_grade5_icon.png";
+
+import artifect_8_grade1_icon from "../../../../assets/union/artifact/artifect_8_grade1_icon.png";
+import artifect_8_grade5_icon from "../../../../assets/union/artifact/artifect_8_grade5_icon.png";
+
+const UnionArtifactIcon = {
+  orange_Mushroom: [artifect_0_grade1_icon, artifect_0_grade5_icon],
+  slime: [artifect_1_grade1_icon, artifect_1_grade5_icon],
+  horny_Mushroom: [artifect_2_grade1_icon, artifect_2_grade5_icon],
+  stump: [artifect_3_grade1_icon, artifect_3_grade5_icon],
+  stone_Golem: [artifect_4_grade1_icon, artifect_4_grade5_icon],
+  balrog: [artifect_5_grade1_icon, artifect_5_grade5_icon],
+  zaqqum: [artifect_6_grade1_icon, artifect_6_grade5_icon],
+  pink_Bean: [artifect_7_grade1_icon, artifect_7_grade5_icon],
+  papulatus: [artifect_8_grade1_icon, artifect_8_grade5_icon],
+};
+export default UnionArtifactIcon;
