+++ src/components/user/union/unionInfo/UnionInfo.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import UnionIcons from "./UnionIcon";
-import { UnionArtifactEffect } from "./UnionArtifactEffect";
+import UnionIcons from "../unionInfo/UnionIcon";
+import { UnionArtifactEffect } from "../unionArtifact/UnionArtifactEffect";
 
 const selectIcon = (level) => {
   let baseLevel, rankCategory;
@@ -102,7 +102,7 @@ const UnionWrap = styled.div`
 `;
 
 const UnionGrade = styled.div`
-  color:rgb(255, 255, 0);
+  color: rgb(255, 255, 0);
   text-shadow: 0px 0px 3px rgb(219, 250, 46);
   font-size: 20px;
   font-family: maple-light;
